third_party/WebKit/LayoutTests/media/video-seeking.html

<!DOCTYPE html>
<title>Test that seeking attribute is true immediately after a seek, goes back to false when seeking completes, and that a "seeked" event is fired for each seek.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var seekedCount = 0;
    var video = document.querySelector("video");
    video.onwaiting = t.step_func(function() {});
    video.onseeked = t.step_func(function() {
        ++seekedCount;

        assert_false(video.seeking);
        assert_equals(video.currentTime, seekedCount * 0.5);

        if (seekedCount == 3) {
            t.done();
            return;
        }

        video.currentTime = (seekedCount + 1) * 0.5;
        assert_true(video.seeking);
    });

    video.src = "content/test.ogv";
    video.currentTime = 0.5;
});
</script>
